[30 mins] step 1: register Junseong UDID on apple developer step 2: add to provisioning profile step 3: publish new build via test flight step 4: chat with Junseong for a few questions
[3 hours]
Get case details screen to properly load options for first suggested case [Done]
Order the properties array by priority also (cases were sorting by priority, but this array was just sorting default)
Need to handle cases where propertyType is N.  These should show up light gray in the picker wheel.  These should not show up in the suggested answers.  When they are selected, the tableview with options should not be visible.
Setting didSelectedEvent on picker wheel to not do anything for N type properties and to show an alertView for info messages
Need to drop any swiped off caseItems in “template mode” from the XML I send when creating the new case
need to remove these items from the picker view
need to make sure these items are not still selected when the user goes to press the update answers button (editing selectedItemForUpdate)—still in progress
renaming the update answers button to “Create Case” when the user is in the mode of having no caseObject ID
Need to remove the suggestedItems that are swiped off from the pickerview scroll wheel [Done]