[2 hours 15 mins]
-  I am unable to start a new case consistently.  When I chose to create new case using a different template or even the same one as before, the app will always give me the previously stored (internally) data.   I see that it sends the request to get a new template but then it ignores the new template that is available in parse.
testing this further, I’ve isolated that it only becomes an issue if I update answers on the form.  If I back out immediately and do the other template, there is no issue.
tracked it down further to root cause.   The code to fetch the ITSMTLObject is executing but it doesn’t appear to be truly refreshing.  In fact, it even reflects data saved to it on the other view controller.  [itsMTLObject fetchInBackgroundWithBlock:^(PFObject *object, NSError *error)
need to develop a strategy for dealing with this one

- I am unable to create a new case.  When I try to create a new case by clicking Update Answers, the app is sending allproperty definitions to the backend.  It should not be sending those.  It should only be sending my answers without <PROPERTY> tags
Removed all property tags from updates except for items in a “newly created properties” index.
Hooked up data for the add new property view controller so it’s adding to the main arrays of case items and properties on the case details screen.
fixed a small bug so the add new property view controller is properly dismissing.
investigating how to properly update the indexing on the pickerview after adding a new case item
investigating a crash regarding the new case item’s selection on the picker view [fixed—need to make sure I am adding to the options string for the new property in the same way as other properties, a semi-colon delimited list instead of an array.]
investigating another crash which occurs when I select multiple answers on the new property UI: <__NSArrayM: 0x7ffb18f4c900> was mutated while being enumerated.’ [fixed. was removing from the array while looping through it..]
[bug] xml writing error when creating xml for case with a new custom property. Likely it has a field or two which are nil and causing exceptions [fixed but XML has error still it seems, email sent]
debugging one more bug on xml writing...'NSInvalidArgumentException', reason: '-[__NSCFNumber _fastCharacterContents]: unrecognized selector sent to instance 0xb000000000000013’ [fixed—I was passing a NSNumber directly for the answerNumber instead of a string]