[Monday February 23] [5 hours 50 minutes] A.replace all "inboundZITSMTL" calls with "submitXML” calls [done] B. work fixing custom answers for template mode 	-fixing a bug where re-entering the custom properties tab after adding 1 custom answer re-sets the first 	custom answer to a type “a” when it should stay as type “custom" 	-in template mode, if the array contains custom..it should update correctly
LOGIC: Any time the user enters the case details page to enter a template, the program will create an array with the count of options in each of the template’s original properties.
If the user goes to edit the template, any tableView selection on the property popup within this original count will be edited back and forth as a value type “a” with a numeric index.  Any value that exceeds this original count will be something the user has added as a custom answer to the template.

This will be marked as value type custom.  Future edits will ensure this is added/removed as answer type custom because it will always be compared to the original count array.  Two or more custom answers may be added to a property and will appear in xml as a semi-colon delimited list alongside normal answers

<ANSWER><A>3;4</A><CUSTOM>Spot;CustomSpot2</CUSTOM></ANSWER>